+++ NEWS.md
@@ -14,6 +14,8 @@ libFirm 1.22.1 (2016-01-07)
 * Improve handling of negative overflow in float to int tarval conversion
 * Improve matching of immediates during instruction selection (amd64)
 * Add peephole optimization 'mov $0, %r' -> 'xorl %r, %r' (amd64)
+* Add peephole optimization 'lea c(%r), %r' -> 'add $c, %r'
+* Add peephole optimization 'lea (%r1, %r2), %r1' -> 'add %r2, %r1' and the commutated case (amd64)
 * Bugfixes
 
 libFirm 1.22.0 (2015-12-31)